+++ docs/changelog.md
@@ -3,3 +3,7 @@
 ## 2023-06-18
 
 - initial creation which is code refactored out of `signal_md` 
+
+## 2024-01-02
+
+- changed `phone-number` or `number` to `mobile` in `people.json`